BotImg/requirements.txt
certifi==2019.11.28
chardet==3.0.4
flickrapi==2.4.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
urllib3==1.25.8